refactor Json for testing in MainBusiness
moved the database connection to Main in try-with-resources block so that it closes connection after execution.  Refactor main to move code for Json testing into a method for readability.  Enable database command code now that the connection is working.  Added granting user to SQL to resove issues of a root password in MySql